Add Two Numbers
create a dummy list and create a curr list to point to dummy. while l1 is not empty or l2 is not empty or carry is there then sum the node values and define carry and value by // and % and create a node with the value and move the linked list pointer and lists pointer. Return dummy.next